+++ src/components/GatewaySelector.tsx
@@ -54,7 +54,7 @@ const GatewaySelectorModal = ({
           gateway,
           label: gateway.settings.label,
           rewardShareRatio: gateway.settings.delegateRewardShareRatio,
-          totalStake: gateway.totalDelegatedStake,
+          totalStake: new mARIOToken(gateway.totalDelegatedStake).toARIO().valueOf(),
           eay: calculateGatewayRewards(
             new mARIOToken(prototocolBalance).toARIO(),
             Object.values(totalGateways).filter((g) => g.status == 'joined')