+++ src/constants.ts
@@ -23,8 +23,11 @@ export const AO_CU_URL = process.env.VITE_AO_CU_URL || 'https://cu.ardrive.io';
 export const DEFAULT_ARWEAVE_PROTOCOL =
   process.env.VITE_GATEWAY_PROTOCOL ?? 'https';
 export const DEFAULT_ARWEAVE_HOST =
-  // process.env.VITE_GATEWAY_HOST ?? 'ar-io.dev';
   process.env.VITE_GATEWAY_HOST ?? 'arweave.net';
+
+export const DEFAULT_ARWEAVE_GQL_ENDPOINT =
+  process.env.VITE_ARWEAVE_GQL_ENDPOINT ??
+  'https://arweave-search.goldsky.com/graphql';
 export const DEFAULT_ARWEAVE_PORT =
   Number(process.env.VITE_GATEWAY_PORT) ?? 443;
 
@@ -59,8 +62,6 @@ export const OPERATOR_EAY_TOOLTIP_FORMULA =
 export const NAME_PASS_THRESHOLD = 0.8;
 export const REFERENCE_GATEWAY_FQDN =
   process.env.VITE_REFERENCE_GATEWAY_FQDN ?? 'arweave.net';
-
-export const GATEWAY_OPERATOR_STAKE_MINIMUM_ARIO = 10000;
 
 export const REDELEGATION_FEE_TOOLTIP_TEXT =
   'Redelegation fees are assessed at 10% per redelegation performed since the last fee reset, up to 60%. Fees are reset when no redelegations are performed in the last 7 days.';